Drop all globbed sass imports
No longer supported now that Steam has moved to Sprockets 3 which is
incompatible with sprockets-sass. An alternative gem is rails-sass,
which first introduced glob imports for sass, but it requires Rails
(surprised?). In the Locomotive world we must avoid Rails dependencies
for Steam since it's shared by both Engine and Wagon, and Wagon sites
are Rack apps, not Rails apps.

Best to avoid glob imports for sass anyway since it relinquishes
control of the cascade ordering...

Should have been included as part of previous commit (cbfea8b)
upgrading to Wagon 2.0.0.rc7, which includes the bump to Steam
1.0.0rc10, which drops sprockets-sass:
https://github.com/locomotivecms/steam/commit/ace09d4163b0fb537b42e14007dc9965a1fb7057#diff-8819e528fc1030c185b10ca998cfd88dL2

Get all that?