npm + webpack for managing dependencies
Taking the plunge...

This was the simplest & most straightforward approach I could come up with to
manage versions and dependencies for front-end components. See cappadona/wagon-
gen-bourbon@0530ae0 for more background on the inherent issues with asset
management in Wagon.

This bumps bourbon to lastest stable, v4.2.6, and now we actually have the
version tracked in the repo.

Will update README with relevant steps for initial setup.

Quick rundown:

* for now Webpack is simply copying src files from modules to public/vendor

* no compiling and public/vendor is not tracked thanks to .gitignore

* provides some sanity for clean @import paths while tracking versions &
  dependencies via package.json