Vendorize Bitters
We'll be using it as a baseline and extending it. This seems to be the best
approach in light of BEM and Bourbon v5.

It will still allow us to track the Bitters version and easily update when
necessary. It has the added benefit of enabling injection of our config in place
of the default Bitters variables, which will not be used, by calling each Bitter
component directly (see base/_base.scss).