New page for events prototype
Went with a new sub-page under news & events landing for now since placing the
prototype on the landing page would require a change to the layouts hierarchy. I
want to avoid that until after merging the `theme-it` branch into dev.

So this first commit simply creates a new page for the prototype. One important
note:

* this page extends layouts/default instead of parent as is the norm for a page
  at the third level

Extending the default layout gives us access to the `main` block in the liquid
templates, which is where we're going to inject the prototype.

The Haml included in this first commit is the direct output of taking the HTML
from last year's Rails prototype and running it through html2haml Heroku app:
http://html2haml.herokuapp.com

In other words, this commit acts as a baseline to better illustrate the
conversion process.