Enable autoprefixer via config/autoprefixer.yml
Awesome how the timing worked out on this...just as we're moving to Bourbon v5
which drops browser prefixes in favor of using a post processor like postcss +
autoprefixer, Didier implements this capability into Steam upstream.

See locomotivecms/wagon#110 and locomotivecms/steam#54.

Still would like to see some improvement on sourcemap capability for the Sass
compilation prior to the handoff to postcss and autoprefixer.