Component for featured via BEM
This one is a bit more rough around the edges than some of the others since it's
using a refill as a placeholder and Refills hasn't yet been updated to for
Bourbon v5 compatibility. In other words, expect these class names to change,
but the layout is coming together.

Pay no attention to events on the home page uncooperating and skipping to the
head of the class...they're up soon.