Rough initial template for Find a Space
Routed to /use/spaces. Was not able to use a regexp as a param for `with_scope`
when dealing with `spaces_types` field due to the many_to_many. Unclear whether
this is a current limitation of locomotive or me ;)

Will investigate further to clean this up. At the very least will move to erb
for this file.