Template for Find a Space
In pure liquid because there's a lot of nesting that Haml will complain about.

This includes a major workaround/hack to build an array of types due to a bug
when using the `with_scope` tag with relationship fields. I spoke to Didier
about this on Gitter and he's going to look at it today since he'd like to
include the bugfix as part of 3.1, which he planned to release today.