Test for page.listed? in nav snippet
This is the default behavior for the nav tag, but I believe you need to do this
explicitly when using a snippet since you're essentially overriding the default
rendering. Makes sense in my head at least ;)

The content type templates for spaces we're already had `listed` set to `false`,
which is what led me here. #241