Update site search to use latest scrape
Going with a new call to the API that will return all items from the given
spider, not specific to a single job. The hope is that I can use the scapinghub
dashboard to remove the results from previous jobs, thereby avoiding duplicates
and the need to continuously have to update to the latest job here. I'm not
confident this will work, but a kid can dream...

Also, had an issue with running job #4 for the spider...scheduled it and after
almost 11 hours it still hadn't run. Cancelled it and then started job #5, which
also sat there waiting for about 10 minutes, so I chatted with Richard in
customer support and he was able to get it going. Sounds like they have a bug in
the new platform which they officially rolled out within the past two weeks.