Overhaul of single space template
Very much a work in progress and there's a lot more that I'd like to do with
this but it's a step in the right direction and should help to elicit
feedback/reactions from our users for this round.

At the very least, there's plenty of DRYing to be done with this heavily
borrowed code from spaces-cards snippet.