Update with_scope tags for staff directory #318
Turns out passing both `false` and `nil` filters will work when serving the site
via Wagon, but not when serving via Engine. The whole point of passing both was
to attempt to appease both truthy/falsy gods, but Engine wins out since Wagon is
only for local development.

Since there are differences in evaluating falsy btw Engine and Wagon, the key is
to test != TRUE and the `ne` operator provided via Steam is our savior...

https://github.com/locomotivecms/steam/blob/2e1b6b821d8e0008dd6b9f616e153b9e9b1e980a/lib/locomotive/steam/adapters/memory/condition.rb#L28

I remember stumbling upon these operators a while back when working on the
content types and templates for spaces. They're sorta buried/not obvious in the
documentation (Click "by a date" under examples heading):
https://locomotive-v3.readme.io/docs/tags#with_scope