Laptop availability for homepage via mannservices
Including the number available for now, but this will need to revisited. The
bibID for the Windows laptops includes a lot of garbage, including:

* batteries

* power cords

* older machines no longer circulated

This explains the incorrect total count of 110 windows laptops. I'll reach out
to Tom when he returns from vacation next week to discuss cataloging options.

Also refactors computing section of available now as follows:

* split out desktops and laptops into separate subheadings (OS entry for each)

* utilize two snippets to DRY it up

* desktops are hardcoded as zero available for now -- up next