Refactor LabStats query & parsing #139
Realized that CIT has already setup their machines to use latest LabStats:

* Room 112 -- where the players dwell -- off the lobby

* Room 220A (2nd floor Mac lab)

* plus the 15 machines outside of the fishbowl on the first floor

I've added these three groups to the Mann Library set within LabStats so they're
exposed via the API.

The logic relies on case/when since Labstats returns the groups as an array with the
group name as a value for the label field instead of a hash using the group name
as a key...that would rock. Oh well...this will do for now.

This covers most of the public desktop machines in the library. Working with
Gabriel to fill the remaining gaps:

* PCs and Macs in Stone Computer Center (aka the farm)

* Macs in McManus Study Room (3rd flr balcony overlooking Dean's Room)

* anything else?