Laptop availability for laptops page
Refactors the availnow-computer snippets so they can be utilized here as well
(gotta keep it DRY).

Introduces a new .page-content--accessory container that renders at the top of
th page on mobile and in the sidebar at larger breakpoints. This container will
render even when empty due to the padding and background color. I have styles in
place to handle this but thanks to Haml's default whitespace handling, the
element is not truly empty.

Next up is to put Haml in its place...