Vendorize semantic-ui component styles via webpack
Using minified css for now since Sass port is not yet ready and at the moment
we're not customizing/overriding semantic-ui as well. Bringing in only the
specific components that are needed across the site. Other component styles are
bundled as needed (e.g. transition.min.css for hoursNav.bundle.js).