Requeat a purchase form validation via webpack
The jQuery plugin `formvalidation` hasn't been published to npm so installing
directly from GitHub repo.

Request a purchase form is dead in the water since we've decided to use the
central CUL form (see #311). We probably also should be using semantic-ui's
native form validation as Tahir and I have discussed in the past:
http://semantic-ui.com/behaviors/form.html

It's even more appealing now with the recent addition of dependent validation
introduced in v2.2:
http://semantic-ui.com/introduction/new.html#dependent-form-validation

In other words, most of this commit is a moot point but going through the
motions for completeness and just in case we do end up using this jQuery
plugin...