jQuery bundled via webpack
Not convinced this is the best approach just yet (global bundle for every page),
but page load is greatly improved over fetching from CDN. The problem is that
semantic-ui has a dependency on jQuery and right now we're using a couple
semantic-ui components for the search form, which is in the header, which means
we need it on every page.

Speaking of the search form, I noticed that it is only rendering properly on the
request a purchase page and I realize now that we need the styles for the
semantic-ui input component globally. That fix up next.