Attempt to demystify webpack config
Provide some clarity as to how extract-text-webpack-plugin is being used to
compile the src sass into single file (public/stylesheets/main.css). WIP.

Disabling these css-loader options is useful in dev, but we'll need to decide
how we want to handle the assets in production to optimize performance. To be
clear, this can all be handled via the webpack config (see 'publicPath' option).

https://github.com/webpack/extract-text-webpack-plugin#api