Use webpack-shell-plugin to remove css bundle
Feels a bit hackish but not sure of a better option at this point. Tried
exporting multiple modules as suggested by Localytics team which also
accomplished the goal (by first writing the bundle to main.css and then over-
writing it via the extract-text-webpack-plugin). Not exactly graceful and also
resulted in a increased compile time. Have this stashed in case I want to
revisit it.

http://eng.localytics.com/faster-sass-builds-with-webpack/#separatingcssfromjavascript