Switch to split gutters for susy layout
Default is to add the gutter each column but with `split`, equal margin is added
before (left) and after (right) each column. This allows us to no longer be
concerned with "edge cases" (i.e. using the `last` keyword), but now we do need
to pay attention to nesting:

http://susydocs.oddbird.net/en/latest/toolkit/#nesting

Overall, this feels like a big win and makes our layout all that more flexible.
Next step is sweep the scss for holdouts of the edge keywords to put this theory
to the test.