Drop sui embed component & use iframe directly
Default template for embed component disables scrolling within iframe:
https://github.com/Semantic-Org/Semantic-UI/blob/2.2.2/dist/components/embed.js#L669

But we scrolling because the content of the iframe changes (expands) as
user interacts with LibCal form.

The main downside to this approach is that we need to set an explicit iframe
height in pixels. Currently using 625px, but might end up reducing it if I can
remove some elements (header, breadcrumb, footer) from iframe content. Nice
thing is that because the modal is responsive this adapts nicely at all
breakpoints.

It does appear that you can provide a custom template for embed component, but
not in the mood to mess with it right now. Can revisit later if necessary:
http://semantic-ui.com/modules/embed.html#/settings