Libcal iframe (#456)
* Proof of concept for LibCal public form via iframe

Using semantic-ui modal and embed components, but most likely ditching embed
since it generates an iframe sans scroll bars.

This is currently hard-coded for Mann 102 only for demo purposes. Will need to
abstract the logic if we decide we're happy with this approach.

* Drop sui embed component & use iframe directly

Default template for embed component disables scrolling within iframe:
https://github.com/Semantic-Org/Semantic-UI/blob/2.2.2/dist/components/embed.js#L669

But we scrolling because the content of the iframe changes (expands) as
user interacts with LibCal form.

The main downside to this approach is that we need to set an explicit iframe
height in pixels. Currently using 625px, but might end up reducing it if I can
remove some elements (header, breadcrumb, footer) from iframe content. Nice
thing is that because the modal is responsive this adapts nicely at all
breakpoints.

It does appear that you can provide a custom template for embed component, but
not in the mood to mess with it right now. Can revisit later if necessary:
http://semantic-ui.com/modules/embed.html#/settings

* Prep handling multiple LibCal widgets on same page

Logic is still hardcoded for Mann 102 only for demo purposes, but this will be
needed when we extend the behavior to Mann 100, Lobby tabling and Stone
Classroom.