Adios booked (#459)
* Strip all (most) references of Booked

Leaving events (Vue.js) related integrations for Tahir to address.

* Use snippet for reservable space button

DRY it up.

* Trigger LibCal widget vs. full via content data

Introduces new field for LibCal group public page URL (aka friendly URL) and
abstracts out the specific implementation used to demo with Mann 102 in #456.

* No need to wrap reserve button within an anchor

This was only causing the browser to append a pound sign to the URL which makes
the history messy. Yes, another approach could have been to use
event.preventDefault(), but this is simpler and cleaner.

P.S. The title attribute originally on the anchor, was moved to the button.

* Clarify differences btw reservation system fields

Make it a point to clearly indicate that two of them apply to LibCal only.