Bite the bullet; track all existing pages
Had been trying to avoid this but enough is enough. This commit ensures every
page is extending the correct layout or parent, and renames the old `section`
layout to `basic-page` -- homage to JCR :)

Add all pages and pass on the content/data for editable_elements (contained in
the YAML header of each file).

There are some exceptions to this where the editable_element was incorrect and I
didn't want to risk losing content again.