Fill-in subpages for Find a Space on 'Use' landing
Extract the array creation originally in `primary-nav` snippet (44a87bfb) into a
snippet of its own so it can be reused here.

The `landing` and `primary-nav` snippets are even more similar after this
commit. I've said it before (c80e7c08) and I'll say it again... these two are
ripe for DRYing up.