Render description on section pages
Initially was thinking we would go out of our way to hide the section pages
(redirects & breadcrumb voodoo), but after implementing most of the IA refactor,
I think there's a cleaner approach that helps to avoid a maintenance headache.

While we aren't linking to the section pages (secondary level) in the megamenu,
we're going to leave them linkable in the breadcrumbs and do a reasonable job
rendering them for any user who wants to browse/explore via the site hierarchy.