Refactor layouts (#463)
* New approach to primary (top-level) page layouts

Two kinds of primary pages:

1. Those that show up in the primary nav and are landing pages (collection of
   links to children pages without any content themselves). These will use
   `primary-landing` layout.

   * Use the Library
   * Get Help
   * News & Events

2. Those that aren't in the primary nav, have content of their own, and aren't
   templates for content types. These will use the `primary` layout.

   * About Us
   * Contact Us

* Convert to straight up liquid -- no Haml

* First go at gallery layout for landing pages

No images yet and no special handling for the problem child (aka Find a Space).

Notice that this new `landing` snippet is very similar to the `primary-nav`
snippet. Would be nice to DRY this up down the road.

* Add placeholder images for landing page

Temporarily using `placeholder-staff.jpg` --need to generalize this-- and
haven't yet implemented the editable file for the `primary-landing` layout. Both
of these up next.

* Generalize placeholder image

Refactor placeholder logic for staff directory while we're at it using the
default filter:
https://locomotive-v3.readme.io/docs/filters#default

* Initial attempt at editable_file for section-photo

This works as expected when serving the site via wagon, but the editable_file
included within the `page.editable_elements` drop always returns `nil` when
serving the site via engine.

Only applied this temporary `section-new` layout to a single page
(/use/technology) for now. Major overhaul to layouts coming up.

* Bite the bullet; track all existing pages

Had been trying to avoid this but enough is enough. This commit ensures every
page is extending the correct layout or parent, and renames the old `section`
layout to `basic-page` -- homage to JCR :)

Add all pages and pass on the content/data for editable_elements (contained in
the YAML header of each file).

There are some exceptions to this where the editable_element was incorrect and I
didn't want to risk losing content again.

* Temporarily comment out editable_file debugging

* Fill-in subpages for Find a Space on 'Use' landing

Extract the array creation originally in `primary-nav` snippet (44a87bfb) into a
snippet of its own so it can be reused here.

The `landing` and `primary-nav` snippets are even more similar after this
commit. I've said it before (c80e7c08) and I'll say it again... these two are
ripe for DRYing up.

* Line-height and margin for landing subpage listing

Improve readability when a list item wraps to multiple lines.

* Unlink secondary pages in megamenu

Now that we have the landing pages in place, we don't want to send users to
secondary pages. Still need to address the breadcrumbs.

* Render description on section pages

Initially was thinking we would go out of our way to hide the section pages
(redirects & breadcrumb voodoo), but after implementing most of the IA refactor,
I think there's a cleaner approach that helps to avoid a maintenance headache.

While we aren't linking to the section pages (secondary level) in the megamenu,
we're going to leave them linkable in the breadcrumbs and do a reasonable job
rendering them for any user who wants to browse/explore via the site hierarchy.

* DRY up `landing` and `primary-nav` snippets

Via `nav-builder` snippet which relies on passing the component name and then
BEM conventions take over from there.

* Rename `primary-nav` snippet to `nav-megamenu`

More descriptive of what it's responsible for and keeps all the nav snippets
together in the directory by using the common prefix.