Delay reveal of megamenu by 200ms & add down caret
This is a very quick & simple first attempt to address the number one issue we
observed in multiple rounds of usability testing where users were inadvertently
triggering the megamenu as the moved the cursor past the top level pages in the
navigation.

This current implementation is problematic on touch devices, so rest assured
this is not the final approach.