Refactor carousel impl; now we have 2 on homepage
Happy happy joy joy!

Includes:

* tweak existing SCSS Webpack loader (extracting to CSS) to target only
  main.scss

* introduce new SCSS loader to handle any Sass imported via JS components
  - no funny business extracting to CSS -- straight up !style!css!sass

* now importing `spotlights` and `meet-experts` components via JS so remove
  leading underscore from filenames and introduce new `carousel` component,
  which both import (via Sass)
  - these components need to import all of the dependent mixin libraries also
    imported in main.scss (breakpoint, susy, bourbon, etc) plus the config
    variables (config, grid)
  - might take this a step further down the road and create two Sass files, one
    that imports the mixin libs and the other that imports the config variables,
    which could then be imported instead...stil undecided

The experts carousel is responsive and shows only 2 slides/experts at the
default mobile first breakpoint. There is a rendering issue at this breakpoint
(the Book a Consultation button is too wide) for the available span, but not
going to worry about it at the moment since there are rumors that this feature
might be removed from the v1 release.