Experts carousel (#488)
* Rename "Meet Our Librarians" to "Meet Our Experts"

Update components and snippets to follow suit. #452

* First go at carousel for meet our experts

Todo:

* Refactor SCSS components to address both homepage carousels

* Consider using CommonsChunkPlugin for Webpack for shared/common code

* Refactor carousel impl; now we have 2 on homepage

Happy happy joy joy!

Includes:

* tweak existing SCSS Webpack loader (extracting to CSS) to target only
  main.scss

* introduce new SCSS loader to handle any Sass imported via JS components
  - no funny business extracting to CSS -- straight up !style!css!sass

* now importing `spotlights` and `meet-experts` components via JS so remove
  leading underscore from filenames and introduce new `carousel` component,
  which both import (via Sass)
  - these components need to import all of the dependent mixin libraries also
    imported in main.scss (breakpoint, susy, bourbon, etc) plus the config
    variables (config, grid)
  - might take this a step further down the road and create two Sass files, one
    that imports the mixin libs and the other that imports the config variables,
    which could then be imported instead...stil undecided

The experts carousel is responsive and shows only 2 slides/experts at the
default mobile first breakpoint. There is a rendering issue at this breakpoint
(the Book a Consultation button is too wide) for the available span, but not
going to worry about it at the moment since there are rumors that this feature
might be removed from the v1 release.

* Remove temp hack from _shame.scss for meet-experts

Temp hack is no more, but in return now we have multiple carousels on the
homepage. You win some, you lose some.

* Consistency in regex for Webpack loaders

* Refactor Sass src to accomodate component imports

I'm liking this...the helpers directory has been riding the bench all this time,
screaming "Put me in, Coach" and this just makes almost too much sense ;)

Includes:

* @import for mixin libraries moved to `helpers/_mixins.scss`

* rips (mostly) Susy related variables out of `layout/_grid.scss` and into
  `_config.scss` where they belong

* simplifies @import's within Webpack imported components (for now that's the
  two homepage carousels -- spotlights & experts)