Set explicit value for checkbox input
For some reason, Locomotive will set a value of 'on' for a checked checkbox,
instead of 1 or true as expected. This results in the user's actual response not
being submitted properly since the underlying field is a boolean.

Don't ask how long it took me to debug this, but once I confirmed what the
submitted value was, a quick search in the repo uncovered
locomotivecms/engine#399. An oldie, but goodie.