Functional multi-select instruction types
Follow-up from previous commit 386de3b which details how multiple values could
be selected but not stored. This workaround involves introduction of a new
field, which means we now have two fields to support multiple values for
instruction types:

1. `instruction_types_managed_list` field (field type: select) to manage the
   options

2. `insruction_types` field (field type: tags) to store the user selected and
   submitted values

This approach does introduce a wrinkle into the patterns we've established for
form-building, but it allows content editors to manage select options from the
back-office instead of locking this up in code. An alternative considered was to
manage the options via `site_metafields`, but this has the disadvantage of
requiring the content editor to go to two disparate sections of the back- office
to accomplish the same task...

Also note the following additional special handling to accommodate this feature:

* element name in the liquid markup for the form has trailing '[]'
  `content[instruction_types][]`

* SUI validation doesn't like this name, so using the `data-validate` attribute
  as the identifier