Final (almost) tweaks to instruction request form (#491)
* Remove "Course Web Page" option from instruct type

Add new boolean field to indicate interest in creation of a LibGuide. As per
conversation with Kelee in #300.

* Set explicit value for checkbox input

For some reason, Locomotive will set a value of 'on' for a checked checkbox,
instead of 1 or true as expected. This results in the user's actual response not
being submitted properly since the underlying field is a boolean.

Don't ask how long it took me to debug this, but once I confirmed what the
submitted value was, a quick search in the repo uncovered
locomotivecms/engine#399. An oldie, but goodie.

* A little less casual for the submit button

As per feedback during staff show 'n tell. Especially important for the
instruction request form, whose primary audience is faculty.

* Enable multiple selection for course instruct type

Kinda. Sort of. Here's the deal:

* the select field type in locomotive does not support multiple values

* so although the UI now allows for multiple selections, only the first one is
  being stored

* if you're thinking we can just change the field type, that would be great, but
  moving away from the `select` field type means we can no longer manage the
  options presented to the user from within the back-office and it would need to
  be handled via code...not ideal

* I've got a few ideas/workarounds for this, which I will pursue next

* one more thing...the delete icon that should appear with each selected option
  is currently not rendering due to the previously mentioned issue with SUI
  v2.2.2 method of porting Font Awesome. I still plan on addressing this at some
  point.

* Functional multi-select instruction types

Follow-up from previous commit 386de3b which details how multiple values could
be selected but not stored. This workaround involves introduction of a new
field, which means we now have two fields to support multiple values for
instruction types:

1. `instruction_types_managed_list` field (field type: select) to manage the
   options

2. `insruction_types` field (field type: tags) to store the user selected and
   submitted values

This approach does introduce a wrinkle into the patterns we've established for
form-building, but it allows content editors to manage select options from the
back-office instead of locking this up in code. An alternative considered was to
manage the options via `site_metafields`, but this has the disadvantage of
requiring the content editor to go to two disparate sections of the back- office
to accomplish the same task...

Also note the following additional special handling to accommodate this feature:

* element name in the liquid markup for the form has trailing '[]'
  `content[instruction_types][]`

* SUI validation doesn't like this name, so using the `data-validate` attribute
  as the identifier