Fix validation bug in consult & instruct forms
This is for server-side validation only, which admittedly should be a rare
occurence since it would only be encountered when JS is disabled.

To retain values from the current submission in progress, the singular version
of the content type's slug should be used.

This is an easily missed detail in the Locomotive guide:
https://locomotive-v3.readme.io/docs/create-a-contact-
form#2-create-a-form

Also, note that when it comes to `select` fields, this server-side validation
will only work when serving the site via engine.