Month/Week toggle now reflects current status
Previously was not toggling active state.

This switching still needs work. Ideally, when going from week to month or vice
versa, the current week/month wouldn't be reset to today. The expected behavior
would be to just switch views and not jump to a different point in time.

Will add a new issue to address this.