Style LibCal monthly hours widget (#519)
* Bundle hours page styles via webpack

No longer imported via main.scss. All possible thanks to 9d1c31e.

* Hours in pure liquid...Haml be gone

* Consistency btw variable & name

Doh...forgot about this when refactoring in 697e275.

* Style the LibCal monthly hours widget

This is mostly good to go, but might replace the full day of the week name with
the 3-character abbreviation via JS due to overlap/crowding at mobile first
(portrait).

* Separate component for LibCal monthly hours widget

* Use variation of closed color for LibCal widget

* Month/Week toggle now reflects current status

Previously was not toggling active state.

This switching still needs work. Ideally, when going from week to month or vice
versa, the current week/month wouldn't be reset to today. The expected behavior
would be to just switch views and not jump to a different point in time.

Will add a new issue to address this.