IA cleanup (#524)
* Move Upcoming Workshops as child of Events

Reflects recent change made by Tobi in back-office. This illuminates the trouble
with tracking pages, even stubs of the pages, in version control.

* Track stub of straggler tutorial pages

Should have been included as part of c572430 while working on #463.

* Update title of basic-page layout #505

So now it's accurately identified in the select list in the back-office.

* When building nav, ensure page is published #510

Previously was only testing that it was listed, but both must be true.