Workshops nav shuffle (#526)
* Remove redirect for workshops after IA shuffle

This is all in an effort to keep Git in sync with back-office changes. Once
again, tracking stubs of all pages in version control = trouble.

Seriously reconsidering this decision. Might be worth going back to ignoring all
pages by default and only explicitly tracking those we are manipulating via
code. And the safest way to do that might be to keep most of the juicy bits in
snippets and this way we're only tracking a handful of includes at the page
level.

* Fix section layout title & reveal in back-office

No longer new and should be selectable from back-office. Continuation of
eac466b.