Adios Haml (almost)
Converts all remaining Haml templates to straight-up Liquid, with a single
exception of `basic-page` layout. The reason being that we still need to rely on
Haml for whitespace control, until Locomotive's fork of Liquid merge's upstream
changes to take advantage of the new whitespace controls. locomotivecms/steam#76

This also includes a tweak to `.page-content` style so it's nested in Susy
layout.