Ignore pages and double-down on snippets (#538)
* Ignore all pages by default #529

Introduces a directory-specific .gitignore for app/views/pages. This should help
prevent headaches trying to keep repo in sync with content changes happening in
back-office. We'll need to edit the list of exceptions as we introduce new
dynamic content.

Default is to ignore all pages. Here are the exceptions to that rule (aka what
will be tracked):

* all layouts (target a sub-dir so future proof)

* all content-type templates (target file name pattern so future-proof)

* 3 top-level pages (makes it easier to develop nav when serving via wagon)

* specific pages (that provide dynamic content)

This initial list of exceptions should look awfully familiar to the pages in the
repo prior to the refactoring (#463) that starting tracking everything:
https://github.com/cul-it/mann-wagon/tree/5f16bb3e10d5730265a7d106f397b2199b96b03a/app/views/pages

* Adios Haml (almost)

Converts all remaining Haml templates to straight-up Liquid, with a single
exception of `basic-page` layout. The reason being that we still need to rely on
Haml for whitespace control, until Locomotive's fork of Liquid merge's upstream
changes to take advantage of the new whitespace controls. locomotivecms/steam#76

This also includes a tweak to `.page-content` style so it's nested in Susy
layout.

* The final piece to the puzzle - lean on snippets

As mentioned in #529, move all custom templating out of pages and into snippets,
so the pages are only left with includes. This will help safeguard against an
end-user accidentally erasing code in the back-office when switching layouts.

https://locomotive-v3.readme.io/docs/layouts#use-layouts