Course reserves (#555)
* Start tracking existing course reserves page

More like tracking it once again ;)

* Initial go at course reserves in Vue (WIP)

* course reserves page includes snippet

* snippet immediately passes off to Vue

* tweak Webpack config to use babel-loader for course-reserves
  -- leaves events Vue app untouched

* table of reserve items treats everything the same at the moment

TODO:

* improve performance of initial load for course list
  -- should look into caching this and potentially even moving it into Liquid
  since I believe the caching within vue-resource would be specific to each
  user/session.

* decorators for items based on item type

* link to Blackboard for eletronic resources

* styles

* Render table only when there are items to display

Plus some styles for course title & instructor.

* Sync course reserves page content from back office

* Stick with kebab-case for asset file names

* New child component for course reserves

Each reserve item for a course is passed to the reserve-item component. There's
a bunch of view logic that needs to happen per item and this component currently
handles that in a single itemLogistics() method.

Includes styling and handling of the following:

* label electronic resources as "Online" & link to Blackboard

* badge available items

* identify location of all physical items not at Mann

More to come...

* Format due dates & indicate overdue items

Moment.js with the big assist.

* Tweak online logic for better consistency

There is some dirty data being returned that requires additional scrubbing. Also
upped the anty on the logical operators.

* Table striping for reserve items

* Title handling when reserve item is an article

Similar to what we're doing on the current site, but now the article title
doesn't take center stage if present.

* Build course list in Liquid = server-side caching

This eliminates the CourseReserves Vue component, since the root vm can now pass
off directly to the ReserveItem component.

I didn't see a better way around the long load times for the course list from
libservices. Using Locomotive's consume tag, we can cache the response (only
applies when the site is hosted via engine). Currently caching for 1 hr.

Looking into vue-resource, it does appear to be possible to cache the response
via interceptors, but this would still be client-side which means any visitor
would experience this initial long load for course listings.
https://github.com/vuejs/vue-resource/issues/252#issuecomment-243611698
https://github.com/vuejs/vue-resource/blob/master/docs/http.md#interceptors