Rough-in the software list table
Proof of concept:

* Fetch raw CSV as string via raw.githubusercontent.com

* Parse and convert to JSON via csvtojson package

* Render as table...no frills

This looks promising. If you checkout this commit, this feature will not be
functional since the CSV currently resides in a private repo and I'm not
including the token. The current plan is to create a new public repo for the
CSV.

I'll most likely move the fetching of the CSV into Liquid to take advantage of
rails caching (following the lead of course-reserves).