Once again switch to fetch data via Loco's consume
The fetching part was simple and was done to take advantage of Rails caching
(caching the response for an hour here as well).

The painful part was returning the multiline response string to its original
form so that it could be passed to Vue. I expect this would have been easier if
Locomotive's fork of Liquid was running the latest improvements to whitespace
control introduced upstream this past summer. See locomotivecms/steam#76