Sticky table headers via stickySort plugin
This was trickier than it appeared. Most of the sticky techniques would result
in the table collapsing when the thead was pulled out of the content flow. This
plugin takes a different approach (adding a wrapper element and copies classes
over). It's not perfect, but it avoids me having to write something from
scratch and gets the job done for now.