Software list (#567)
* Start tracking existing software page

* Remove straggler editable elements

Leftover from when the page was extending the section layout, potentially
inadvertently by extending the parent by default.

* Rough-in the software list table

Proof of concept:

* Fetch raw CSV as string via raw.githubusercontent.com

* Parse and convert to JSON via csvtojson package

* Render as table...no frills

This looks promising. If you checkout this commit, this feature will not be
functional since the CSV currently resides in a private repo and I'm not
including the token. The current plan is to create a new public repo for the
CSV.

I'll most likely move the fetching of the CSV into Liquid to take advantage of
rails caching (following the lead of course-reserves).

* Switch to new public GitHub repo for csv fetching

Public = no token :)

* Once again switch to fetch data via Loco's consume

The fetching part was simple and was done to take advantage of Rails caching
(caching the response for an hour here as well).

The painful part was returning the multiline response string to its original
form so that it could be passed to Vue. I expect this would have been easier if
Locomotive's fork of Liquid was running the latest improvements to whitespace
control introduced upstream this past summer. See locomotivecms/steam#76

* Comment to explain empty modules in webpack config

Should have included this as part of 39f107c.

* Style the software table

Only thing left on the list is a fixed header...

* Sticky table headers via stickySort plugin

This was trickier than it appeared. Most of the sticky techniques would result
in the table collapsing when the thead was pulled out of the content flow. This
plugin takes a different approach (adding a wrapper element and copies classes
over). It's not perfect, but it avoids me having to write something from
scratch and gets the job done for now.