Initial go at Google CSE #557 #509
Using the `<gcse:searchresults-only>` component and sticking with our
own/existing search box for now. We can revisit this later if we end up wanting
to get fancy with autocomplete, etc.

Some quick style overrides to make this fit in some more. Current indexing of
dev site is limited and right away I'm realizing that we need to update the
`<title>` value for our pages. That's next.